// src/app/models/product.model.ts
export class Product {
  id: string | null = null;
  name: string | null = null;
  description: string | null = null;
  price: number | null = null;
  category: string | null = null;
  brand: string | null = null;
  stock: number | null = null;
  imageUrl: string | null = null;
  rating: number | null = null;
  enabled: boolean | null = null;
}